refactor(load-tests): separate payloads from submitter pacing
Make chain prep a Payload concern, move TxType wiring into WorkloadGenerator,
and split the monolithic load runner into funding and pacing modules.
