+++ crates/infra/load-tests/src/config/real_token.rs
@@ -6,7 +6,7 @@ use super::{
     test_config::{TxTypeConfig, WeightedTxType},
 };
 use crate::{
-    runner::{RealTokenAcquisition, RealTokenPairTokenSetup, RealTokenSetup},
+    workload::{RealTokenAcquisition, RealTokenPairTokenSetup, RealTokenSetup},
     utils::{BaselineError, Result},
 };
 
@@ -257,7 +257,7 @@ mod tests {
     use alloy_primitives::{Address, U256};
 
     use super::super::test_config::TestConfig;
-    use crate::runner::RealTokenAcquisition;
+    use crate::workload::RealTokenAcquisition;
 
     #[test]
     fn parse_real_token_setup_for_random_direction_swap_parity() {
