+++ crates/infra/load-tests/src/lib.rs
@@ -33,10 +33,11 @@ pub use metrics::{
 
 mod workload;
 pub use workload::{
-    AccountPool, AerodromeClPayload, B20TransferPayload, CalldataPayload,
-    Erc20Payload, FundedAccount, KeyStream, OsakaPayload, Payload, PrecompileLooper,
-    PrecompilePayload, SeededRng, StoragePayload, TransferPayload, UniswapV3Payload,
-    WorkloadGenerator, parse_precompile_id,
+    AccountPool, AerodromeClPayload, B20TransferPayload, CalldataPayload, ChainPrepContext,
+    ChainPrepOutputs, Erc20Payload, FundedAccount, KeyStream, OsakaPayload, Payload,
+    PrecompileLooper, PrecompilePayload, RealTokenAcquisition, RealTokenPairTokenSetup,
+    RealTokenRecoverySummary, RealTokenSetup, SeededRng, StoragePayload, TransferPayload,
+    UniswapV3Payload, WorkloadGenerator, parse_precompile_id, recover_real_tokens,
 };
 
 mod runner;
@@ -45,8 +46,7 @@ pub use runner::{
     DEFAULT_MAX_GAS_PRICE, DisplaySnapshot, FlashblockInclusion, FlashblockWatcher, LoadConfig,
     LoadRunner, LoadTestDisplay, MAX_FEE_BASE_FEE_MULTIPLIER, MAX_SENDER_WORKER_COUNT,
     MAX_SIGNER_WORKER_COUNT, PipelineQueue, PipelineStartConfig, PreparedBatch,
-    PreparedTransaction, QueuedSubmitFailures, RealTokenAcquisition, RealTokenPairTokenSetup,
-    RealTokenRecoverySummary, RealTokenSetup, ResultsTracker, SENDER_WORKERS_PER_RPC,
+    PreparedTransaction, QueuedSubmitFailures, ResultsTracker, SENDER_WORKERS_PER_RPC,
     SIGNER_WORKERS_PER_RPC, SUBMIT_BATCH_QUEUE_BUFFER, SUBMIT_MAX_ATTEMPTS, SenderContext,
     SentTransaction, SignedBatch, SignedTransaction, SignerContext, SubmissionPipeline,
     SubmitEvent, TxConfig, TxType,
