+++ crates/infra/load-tests/src/runner/funding.rs
@@ -1,0 +1,964 @@
+//! Account funding, draining, txpool clearing, and swap-token mint helpers.
+
+use std::{
+    collections::{HashMap, HashSet},
+    sync::Arc,
+    time::{Duration, Instant},
+};
+
+use alloy_network::{Ethereum, EthereumWallet, TransactionBuilder};
+use alloy_primitives::{Address, Bytes, U256, utils::format_ether};
+use alloy_provider::{Provider, RootProvider};
+use alloy_rpc_types::{BlockNumberOrTag, TransactionRequest};
+use alloy_signer_local::PrivateKeySigner;
+use alloy_sol_types::{SolCall, sol};
+use base_tx_manager::NonceManager;
+use futures::{StreamExt, stream};
+use indicatif::{ProgressBar, ProgressStyle};
+use tracing::{debug, error, info, instrument, warn};
+
+use super::{LoadRunner, SubmissionPipeline, TxType, load_runner::NONCE_RPC_TIMEOUT};
+use crate::{
+    BaselineError, Result,
+    rpc::{BaseFeeExt, QueryProvider, RpcResultExt, TxpoolAdminClient, create_wallet_provider},
+    workload::{
+        await_token_balances, encode_erc20_balance_of, prep_submission_max_fee,
+    },
+};
+
+/// Maximum number of concurrent RPC requests during funding/draining operations.
+pub(super) const FUNDING_CONCURRENCY: usize = 32;
+
+const TXPOOL_CLEAR_CONCURRENCY: usize = 64;
+
+impl LoadRunner {
+    /// Funds all accounts from a funding key up to the specified amount.
+    pub async fn fund_accounts(
+        &mut self,
+        funding_key: PrivateKeySigner,
+        amount_per_account: U256,
+    ) -> Result<()> {
+        let total_accounts = self.accounts.len();
+        let client = self.client.clone();
+        let primary_submission_rpc = self.config.primary_submission_rpc().clone();
+        let chain_id = self.config.chain_id;
+        let max_gas_price = self.config.max_gas_price;
+
+        let pb_check = self.progress_bar(total_accounts as u64, "Checking balances");
+
+        // Phase 1: Parallel balance + nonce queries.
+        let addresses: Vec<(Address, usize)> =
+            self.accounts.accounts().iter().enumerate().map(|(i, a)| (a.address, i)).collect();
+
+        let balance_futs: Vec<_> = addresses
+            .iter()
+            .map(|&(addr, idx)| {
+                let client = client.clone();
+                async move {
+                    let balance = client.get_balance(addr).await.rpc("get balance")?;
+                    let nonce =
+                        client.get_transaction_count(addr).await.rpc("get transaction count")?;
+                    Ok::<_, BaselineError>((addr, idx, balance, nonce))
+                }
+            })
+            .collect();
+
+        let results: Vec<_> = stream::iter(balance_futs)
+            .buffer_unordered(FUNDING_CONCURRENCY)
+            .inspect(|_| pb_check.inc(1))
+            .collect()
+            .await;
+        pb_check.finish_and_clear();
+
+        let mut accounts_to_fund = Vec::new();
+        for result in results {
+            let (addr, idx, balance, nonce) = result?;
+            let account = &mut self.accounts.accounts_mut()[idx];
+            account.balance = balance;
+            account.nonce = nonce;
+
+            if balance < amount_per_account {
+                let deficit = amount_per_account.saturating_sub(balance);
+                accounts_to_fund.push((addr, deficit));
+            } else {
+                debug!(address = %addr, balance = %balance, "account already funded");
+            }
+        }
+
+        if accounts_to_fund.is_empty() {
+            info!("all accounts already have sufficient balance, skipping funding");
+            return Ok(());
+        }
+
+        let funder_address = funding_key.address();
+        let wallet = EthereumWallet::from(funding_key);
+        let funder_provider =
+            Arc::new(create_wallet_provider(primary_submission_rpc.clone(), wallet));
+
+        let base_fee = client.get_base_fee().await?;
+        let max_priority_fee = (base_fee / 10).max(1);
+        let max_fee =
+            prep_submission_max_fee(base_fee, max_priority_fee, max_gas_price);
+
+        // Phase 2: Early balance validation — abort before sending any TXs if
+        // the funder cannot cover the total cost.
+        let total_deficit: U256 = accounts_to_fund
+            .iter()
+            .map(|(_, deficit)| *deficit)
+            .fold(U256::ZERO, |a, b| a.saturating_add(b));
+        let gas_cost_per_tx = U256::from(21_000u64).saturating_mul(U256::from(max_fee));
+        let total_gas_cost = gas_cost_per_tx.saturating_mul(U256::from(accounts_to_fund.len()));
+        let total_needed = total_deficit.saturating_add(total_gas_cost);
+
+        let funder_balance = client.get_balance(funder_address).await.rpc("get balance")?;
+
+        if funder_balance < total_needed {
+            let shortfall = total_needed.saturating_sub(funder_balance);
+            return Err(BaselineError::Transaction(format!(
+                "funder {} has insufficient balance: has {} ETH, needs {} ETH (deficit {} ETH + gas {} ETH), shortfall {} ETH",
+                funder_address,
+                format_ether(funder_balance),
+                format_ether(total_needed),
+                format_ether(total_deficit),
+                format_ether(total_gas_cost),
+                format_ether(shortfall),
+            )));
+        }
+
+        let start_nonce = funder_provider
+            .get_transaction_count(funder_address)
+            .pending()
+            .await
+            .rpc("get pending transaction count")?;
+
+        info!(
+            from = %funder_address,
+            amount = %amount_per_account,
+            accounts_needing_funds = accounts_to_fund.len(),
+            funder_balance = %format_ether(funder_balance),
+            total_needed = %format_ether(total_needed),
+            "funding accounts"
+        );
+
+        let replacement_max_fee = max_fee.saturating_mul(3);
+        let replacement_priority_fee = max_priority_fee.saturating_mul(3);
+
+        // Phase 3+4: Send funding TXs in batches and confirm each batch before
+        // sending the next. This avoids overwhelming the txpool's per-sender limit.
+        let txs: Vec<(TransactionRequest, Address, U256, u64)> = accounts_to_fund
+            .iter()
+            .enumerate()
+            .map(|(i, &(address, deficit))| {
+                let nonce = start_nonce
+                    .checked_add(u64::try_from(i).expect("account index exceeds u64"))
+                    .expect("nonce overflow");
+                let tx = TransactionRequest::default()
+                    .with_to(address)
+                    .with_value(deficit)
+                    .with_nonce(nonce)
+                    .with_chain_id(chain_id)
+                    .with_gas_limit(21_000)
+                    .with_max_fee_per_gas(max_fee)
+                    .with_max_priority_fee_per_gas(max_priority_fee);
+                (tx, address, deficit, nonce)
+            })
+            .collect();
+
+        let total_txs = txs.len() as u64;
+        let pb_fund = self.progress_bar(total_txs, "Funding accounts");
+        let mut txs_remaining = txs.into_iter().peekable();
+        while txs_remaining.peek().is_some() {
+            let batch: Vec<_> =
+                txs_remaining.by_ref().take(self.config.max_in_flight_per_sender as usize).collect();
+            let mut batch_pending: Vec<Address> = Vec::with_capacity(batch.len());
+            let mut retries: Vec<(Address, U256, u64)> = Vec::new();
+            let mut fatal_errors: Vec<String> = Vec::new();
+
+            let send_futs = batch.into_iter().map(|(tx, address, deficit, nonce)| {
+                let provider = Arc::clone(&funder_provider);
+                async move {
+                    let result = provider.send_transaction(tx).await;
+                    (result, address, deficit, nonce)
+                }
+            });
+
+            let mut send_stream =
+                stream::iter(send_futs).buffer_unordered(self.config.max_in_flight_per_sender as usize);
+
+            let mut nonce_refresh_needed: Vec<(Address, U256)> = Vec::new();
+
+            while let Some((result, address, deficit, nonce)) = send_stream.next().await {
+                match result {
+                    Ok(pending) => {
+                        let tx_hash = *pending.tx_hash();
+                        debug!(to = %address, deficit = %deficit, nonce, tx_hash = %tx_hash, "funding tx sent");
+                        batch_pending.push(address);
+                    }
+                    Err(e) => {
+                        let error_str = e.to_string();
+                        if error_str.contains("already known")
+                            || error_str.contains("replacement transaction underpriced")
+                        {
+                            retries.push((address, deficit, nonce));
+                        } else if error_str.contains("nonce too low") {
+                            info!(to = %address, nonce, "nonce too low, will refresh and retry");
+                            nonce_refresh_needed.push((address, deficit));
+                        } else {
+                            error!(to = %address, error = %e, "failed to fund account");
+                            fatal_errors.push(format!("failed to fund {address}: {e}"));
+                        }
+                    }
+                }
+            }
+
+            if !fatal_errors.is_empty() {
+                pb_fund.finish_and_clear();
+                return Err(BaselineError::Transaction(format!(
+                    "{} funding tx(s) failed: {}",
+                    fatal_errors.len(),
+                    fatal_errors.join("; "),
+                )));
+            }
+
+            if !retries.is_empty() {
+                let retry_futs = retries.into_iter().map(|(address, deficit, nonce)| {
+                    let provider = Arc::clone(&funder_provider);
+                    async move {
+                        let replacement = TransactionRequest::default()
+                            .with_to(address)
+                            .with_value(deficit)
+                            .with_nonce(nonce)
+                            .with_chain_id(chain_id)
+                            .with_gas_limit(21_000)
+                            .with_max_fee_per_gas(replacement_max_fee)
+                            .with_max_priority_fee_per_gas(replacement_priority_fee);
+                        let result = provider.send_transaction(replacement).await;
+                        (result, address, nonce)
+                    }
+                });
+
+                let mut retry_stream =
+                    stream::iter(retry_futs).buffer_unordered(self.config.max_in_flight_per_sender as usize);
+
+                while let Some((result, address, nonce)) = retry_stream.next().await {
+                    match result {
+                        Ok(pending) => {
+                            let tx_hash = *pending.tx_hash();
+                            info!(to = %address, nonce, tx_hash = %tx_hash, "replacement funding tx sent");
+                            batch_pending.push(address);
+                        }
+                        Err(replace_err) => {
+                            warn!(to = %address, nonce, error = %replace_err, "replacement tx also failed, proceeding");
+                        }
+                    }
+                }
+            }
+
+            Self::await_balances(&client, &mut batch_pending, amount_per_account, &pb_fund).await?;
+
+            if !nonce_refresh_needed.is_empty() {
+                let fresh_nonce = funder_provider
+                    .get_transaction_count(funder_address)
+                    .pending()
+                    .await
+                    .rpc("get pending transaction count")?;
+
+                info!(
+                    count = nonce_refresh_needed.len(),
+                    fresh_nonce, "retrying funding txs with refreshed nonce"
+                );
+
+                let nonce_retry_futs =
+                    nonce_refresh_needed.into_iter().enumerate().map(|(i, (address, deficit))| {
+                        let provider = Arc::clone(&funder_provider);
+                        let retry_nonce = fresh_nonce + i as u64;
+                        async move {
+                            let tx = TransactionRequest::default()
+                                .with_to(address)
+                                .with_value(deficit)
+                                .with_nonce(retry_nonce)
+                                .with_chain_id(chain_id)
+                                .with_gas_limit(21_000)
+                                .with_max_fee_per_gas(max_fee)
+                                .with_max_priority_fee_per_gas(max_priority_fee);
+                            let result = provider.send_transaction(tx).await;
+                            (result, address, retry_nonce)
+                        }
+                    });
+
+                let mut nonce_retry_stream =
+                    stream::iter(nonce_retry_futs).buffered(self.config.max_in_flight_per_sender as usize);
+
+                let mut nonce_retry_pending: Vec<Address> = Vec::new();
+                while let Some((result, address, retry_nonce)) = nonce_retry_stream.next().await {
+                    match result {
+                        Ok(pending) => {
+                            let tx_hash = *pending.tx_hash();
+                            info!(to = %address, nonce = retry_nonce, tx_hash = %tx_hash, "nonce-refreshed funding tx sent");
+                            nonce_retry_pending.push(address);
+                        }
+                        Err(retry_err) => {
+                            warn!(to = %address, nonce = retry_nonce, error = %retry_err, "nonce-refreshed retry also failed, proceeding");
+                        }
+                    }
+                }
+
+                Self::await_balances(
+                    &client,
+                    &mut nonce_retry_pending,
+                    amount_per_account,
+                    &pb_fund,
+                )
+                .await?;
+            }
+        }
+        pb_fund.finish_and_clear();
+
+        // Phase 5: Parallel post-funding state refresh.
+        let pb_refresh = self.progress_bar(total_accounts as u64, "Refreshing account state");
+        let refresh_futs: Vec<_> = self
+            .accounts
+            .accounts()
+            .iter()
+            .map(|a| {
+                let client = client.clone();
+                let addr = a.address;
+                async move {
+                    let balance = client.get_balance(addr).await.rpc("get balance")?;
+                    let nonce =
+                        client.get_transaction_count(addr).await.rpc("get transaction count")?;
+                    Ok::<_, BaselineError>((addr, balance, nonce))
+                }
+            })
+            .collect();
+
+        let refresh_results: Vec<_> = stream::iter(refresh_futs)
+            .buffer_unordered(FUNDING_CONCURRENCY)
+            .inspect(|_| pb_refresh.inc(1))
+            .collect()
+            .await;
+        pb_refresh.finish_and_clear();
+
+        let addr_to_idx: HashMap<Address, usize> =
+            self.accounts.accounts().iter().enumerate().map(|(i, a)| (a.address, i)).collect();
+
+        let refresh_provider = RootProvider::<Ethereum>::new_http(self.config.query_rpc.clone());
+
+        for result in refresh_results {
+            let (addr, balance, account_nonce) = result?;
+            let idx = addr_to_idx[&addr];
+            let account = &mut self.accounts.accounts_mut()[idx];
+            account.balance = balance;
+            account.nonce = account_nonce;
+
+            let nonce_manager =
+                NonceManager::new(refresh_provider.clone(), addr, NONCE_RPC_TIMEOUT)
+                    .with_pending_tag();
+            Arc::make_mut(&mut self.nonce_managers).insert(addr, nonce_manager);
+
+            debug!(address = %addr, balance = %balance, nonce = account_nonce, "account state refreshed");
+        }
+
+        info!(funded = accounts_to_fund.len(), "funding complete");
+        Ok(())
+    }
+
+    /// Collects unique token addresses from configured swap transaction types.
+    pub fn collect_swap_tokens(&self) -> Vec<Address> {
+        let mut tokens = HashSet::new();
+        for tx_config in &self.config.transactions {
+            match &tx_config.tx_type {
+                TxType::UniswapV3 { token_in, token_out, .. }
+                | TxType::AerodromeCl { token_in, token_out, .. } => {
+                    tokens.insert(*token_in);
+                    tokens.insert(*token_out);
+                }
+                TxType::Transfer
+                | TxType::Calldata { .. }
+                | TxType::Erc20 { .. }
+                | TxType::Storage { .. }
+                | TxType::B20
+                | TxType::Precompile { .. }
+                | TxType::Osaka { .. } => {}
+            }
+        }
+        tokens.into_iter().collect()
+    }
+
+    /// Collects unique router addresses from configured swap transaction types.
+    pub fn collect_swap_routers(&self) -> Vec<Address> {
+        let mut routers = HashSet::new();
+        for tx_config in &self.config.transactions {
+            match &tx_config.tx_type {
+                TxType::UniswapV3 { router, .. } | TxType::AerodromeCl { router, .. } => {
+                    routers.insert(*router);
+                }
+                TxType::Transfer
+                | TxType::Calldata { .. }
+                | TxType::Erc20 { .. }
+                | TxType::Storage { .. }
+                | TxType::B20
+                | TxType::Precompile { .. }
+                | TxType::Osaka { .. } => {}
+            }
+        }
+        routers.into_iter().collect()
+    }
+
+    /// Clears pending transactions from all configured txpool nodes for every test sender.
+    #[instrument(skip(self), fields(nodes = self.config.txpool_nodes.len(), accounts = self.accounts.len()))]
+    pub async fn clear_txpools(&self) -> Result<u64> {
+        if self.config.txpool_nodes.is_empty() {
+            return Ok(0);
+        }
+
+        info!(
+            nodes = self.config.txpool_nodes.len(),
+            accounts = self.accounts.len(),
+            "clearing txpool sender transactions"
+        );
+
+        let clients: Vec<_> = self
+            .config
+            .txpool_nodes
+            .iter()
+            .cloned()
+            .map(|node| {
+                let client = TxpoolAdminClient::new(node.clone())?;
+                Ok::<_, BaselineError>((node, client))
+            })
+            .collect::<Result<_>>()?;
+        let addresses: Vec<_> =
+            self.accounts.accounts().iter().map(|account| account.address).collect();
+        let requests: Vec<_> = clients
+            .iter()
+            .flat_map(|(node, client)| {
+                addresses
+                    .iter()
+                    .copied()
+                    .map(move |address| (node.clone(), client.clone(), address))
+            })
+            .collect();
+
+        let clear_results: Vec<_> =
+            stream::iter(requests.into_iter().map(|(node, client, address)| async move {
+                let removed = client.drop_sender_transactions(address).await.map_err(|e| {
+                    BaselineError::Rpc(format!(
+                        "failed to clear txpool node {node} for sender {address}: {e}"
+                    ))
+                })?;
+                Ok::<_, BaselineError>((node, removed.len() as u64))
+            }))
+            .buffer_unordered(TXPOOL_CLEAR_CONCURRENCY)
+            .collect()
+            .await;
+
+        let mut removed_by_node: HashMap<url::Url, u64> = HashMap::new();
+        for result in clear_results {
+            let (node, removed) = result?;
+            removed_by_node
+                .entry(node)
+                .and_modify(|total| *total = total.saturating_add(removed))
+                .or_insert(removed);
+        }
+
+        let mut removed_total = 0u64;
+        for node in &self.config.txpool_nodes {
+            let removed_for_node = removed_by_node.get(node).copied().unwrap_or(0);
+            removed_total = removed_total.saturating_add(removed_for_node);
+            info!(
+                node = %node,
+                removed = removed_for_node,
+                "cleared txpool sender transactions from node"
+            );
+        }
+
+        info!(removed = removed_total, "txpool clearing complete");
+        Ok(removed_total)
+    }
+
+    /// Mints swap tokens to all sender accounts.
+    ///
+    /// Scans the configured transaction types for token addresses, then mints
+    /// `amount_per_token` of each token to every sender that has insufficient balance.
+    /// Skips accounts that already have enough tokens. Requires tokens that expose
+    /// a public `mint(address,uint256)` function (e.g., `FreeTransferERC20`).
+    #[instrument(skip(self, funding_key), fields(accounts = self.accounts.len()))]
+    pub async fn setup_swap_tokens(
+        &self,
+        funding_key: PrivateKeySigner,
+        amount_per_token: U256,
+    ) -> Result<()> {
+        let tokens = self.collect_swap_tokens();
+        if tokens.is_empty() {
+            debug!("no swap tokens configured, skipping token setup");
+            return Ok(());
+        }
+
+        let sender_addresses: Vec<Address> =
+            self.accounts.accounts().iter().map(|a| a.address).collect();
+        let token_count = tokens.len();
+        let total_pairs = token_count * sender_addresses.len();
+
+        // Phase 1: Check existing token balances for all (token, sender) pairs.
+        let pb_check = self.progress_bar(total_pairs as u64, "Checking token balances");
+        let client = &self.client;
+
+        let balance_futs: Vec<_> = tokens
+            .iter()
+            .flat_map(|&token| {
+                sender_addresses.iter().map(move |&sender| {
+                    let client = client.clone();
+                    let call_data = encode_erc20_balance_of(sender);
+                    async move {
+                        let result = client
+                            .call(
+                                TransactionRequest::default()
+                                    .with_to(token)
+                                    .with_input(call_data)
+                                    .into(),
+                            )
+                            .await
+                            .rpc("eth_call")
+                            .map(|bytes| U256::from_be_slice(bytes.as_ref()))
+                            .unwrap_or(U256::ZERO);
+                        (token, sender, result)
+                    }
+                })
+            })
+            .collect();
+
+        let balance_results: Vec<_> = stream::iter(balance_futs)
+            .buffer_unordered(FUNDING_CONCURRENCY)
+            .inspect(|_| pb_check.inc(1))
+            .collect()
+            .await;
+        pb_check.finish_and_clear();
+
+        // Filter to only (token, sender) pairs that need funding.
+        let mut transfers_needed: Vec<(Address, Address)> = Vec::new();
+        let mut already_funded = 0usize;
+        for (token, sender, balance) in balance_results {
+            if balance < amount_per_token {
+                transfers_needed.push((token, sender));
+            } else {
+                already_funded += 1;
+                debug!(token = %token, sender = %sender, balance = %balance, "account already has sufficient tokens");
+            }
+        }
+
+        if transfers_needed.is_empty() {
+            info!(
+                tokens = token_count,
+                accounts = sender_addresses.len(),
+                "all accounts already have sufficient token balances, skipping distribution"
+            );
+            return Ok(());
+        }
+
+        info!(
+            transfers_needed = transfers_needed.len(),
+            already_funded = already_funded,
+            tokens = token_count,
+            accounts = sender_addresses.len(),
+            "distributing swap tokens"
+        );
+
+        // Phase 2: Setup for transfers.
+        let funder_address = funding_key.address();
+        let wallet = EthereumWallet::from(funding_key);
+        let funder_provider =
+            Arc::new(create_wallet_provider(self.config.primary_submission_rpc().clone(), wallet));
+        let chain_id = self.config.chain_id;
+        let max_gas_price = self.config.max_gas_price;
+
+        let base_fee = self.client.get_base_fee().await?;
+        let max_priority_fee = (base_fee / 10).max(1);
+        let max_fee =
+            prep_submission_max_fee(base_fee, max_priority_fee, max_gas_price);
+
+        // Pre-flight balance check — abort before sending any TXs if the funder
+        // cannot cover the total gas cost for needed token transfers.
+        let gas_cost_per_tx = U256::from(65_000u64).saturating_mul(U256::from(max_fee));
+        let total_gas_cost = gas_cost_per_tx.saturating_mul(U256::from(transfers_needed.len()));
+        let funder_balance = self.client.get_balance(funder_address).await.rpc("get balance")?;
+
+        if funder_balance < total_gas_cost {
+            let shortfall = total_gas_cost.saturating_sub(funder_balance);
+            return Err(BaselineError::Transaction(format!(
+                "funder {} has insufficient balance for token distribution: has {} ETH, needs {} ETH (gas for {} txs), shortfall {} ETH",
+                funder_address,
+                format_ether(funder_balance),
+                format_ether(total_gas_cost),
+                transfers_needed.len(),
+                format_ether(shortfall),
+            )));
+        }
+
+        let mut nonce = funder_provider
+            .get_transaction_count(funder_address)
+            .pending()
+            .await
+            .rpc("get pending transaction count")?;
+
+        // Phase 3: Execute transfers for accounts that need tokens.
+        let pb = self.progress_bar(transfers_needed.len() as u64, "Minting tokens");
+        let mut failed_count: usize = 0;
+
+        let txs: Vec<(TransactionRequest, Address, Address)> = transfers_needed
+            .into_iter()
+            .map(|(token, sender)| {
+                let mint_data = Self::encode_erc20_mint(sender, amount_per_token);
+                let tx = TransactionRequest::default()
+                    .with_to(token)
+                    .with_input(mint_data)
+                    .with_nonce(nonce)
+                    .with_chain_id(chain_id)
+                    .with_gas_limit(65_000)
+                    .with_max_fee_per_gas(max_fee)
+                    .with_max_priority_fee_per_gas(max_priority_fee);
+                nonce += 1;
+                (tx, token, sender)
+            })
+            .collect();
+
+        let total_txs = txs.len();
+        let mut txs_remaining = txs.into_iter().peekable();
+        while txs_remaining.peek().is_some() {
+            let batch: Vec<_> =
+                txs_remaining.by_ref().take(self.config.max_in_flight_per_sender as usize).collect();
+            let mut pending_txs: Vec<(Address, Address)> = Vec::new();
+
+            let send_futs = batch.into_iter().map(|(tx, token, sender)| {
+                let provider = Arc::clone(&funder_provider);
+                async move {
+                    let result = provider.send_transaction(tx).await;
+                    (result, token, sender)
+                }
+            });
+
+            let mut send_stream =
+                stream::iter(send_futs).buffer_unordered(self.config.max_in_flight_per_sender as usize);
+
+            while let Some((result, token, sender)) = send_stream.next().await {
+                match result {
+                    Ok(pending) => {
+                        let tx_hash = *pending.tx_hash();
+                        debug!(token = %token, to = %sender, tx_hash = %tx_hash, "token mint sent");
+                        pending_txs.push((token, sender));
+                    }
+                    Err(e) => {
+                        warn!(token = %token, to = %sender, error = %e, "token mint failed");
+                        failed_count += 1;
+                    }
+                }
+            }
+
+            await_token_balances(&self.client, &mut pending_txs, amount_per_token, &pb)
+                .await?;
+        }
+
+        pb.finish_and_clear();
+
+        if failed_count > 0 {
+            return Err(BaselineError::Transaction(format!(
+                "{failed_count}/{total_txs} token mints failed — senders with missing tokens will revert on swap"
+            )));
+        }
+
+        info!(
+            tokens = token_count,
+            transfers = total_txs,
+            skipped = already_funded,
+            "swap token setup complete"
+        );
+        Ok(())
+    }
+
+    fn encode_erc20_mint(to: Address, amount: U256) -> Bytes {
+        sol! {
+            function mint(address to, uint256 amount) external;
+        }
+        Bytes::from(mintCall { to, amount }.abi_encode())
+    }
+
+    /// Drains all test account balances back to the funder address.
+    ///
+    /// Each account sends its entire balance minus gas costs back to the funder.
+    /// Transactions that fail (e.g. zero balance) are skipped with a warning.
+    pub async fn drain_accounts(&self, funding_key: PrivateKeySigner) -> Result<U256> {
+        let funder_address = funding_key.address();
+        let client = self.client.clone();
+        let primary_submission_rpc = self.config.primary_submission_rpc().clone();
+        let chain_id = self.config.chain_id;
+
+        let base_fee = client.get_base_fee().await?;
+        let max_priority_fee = (base_fee / 10).max(1);
+        let max_fee = SubmissionPipeline::submission_max_fee(
+            base_fee,
+            max_priority_fee,
+            self.config.max_gas_price,
+        );
+        let drain_gas_limit = 21_000u128;
+        // L1 data fee on Base can be significant (0.0001-0.001 ETH depending on L1 gas prices).
+        // Use 0.001 ETH (1e15 wei) buffer to be safe. We may leave dust in accounts.
+        let l1_fee_buffer = 1_000_000_000_000_000u128;
+        let drain_gas_cost = U256::from(drain_gas_limit * max_fee + l1_fee_buffer);
+
+        let total_accounts = self.accounts.len();
+        let pb_drain = self.progress_bar(total_accounts as u64, "Draining accounts");
+
+        // Each account has its own signer, so drains are fully independent.
+        let account_data: Vec<_> =
+            self.accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
+
+        let drain_futs: Vec<_> = account_data
+            .into_iter()
+            .map(|(address, signer)| {
+                let client = client.clone();
+                let primary_submission_rpc = primary_submission_rpc.clone();
+                async move {
+                    let balance = client
+                        .get_balance(address)
+                        .block_id(BlockNumberOrTag::Pending.into())
+                        .await
+                        .rpc("get pending balance")?;
+                    if balance <= drain_gas_cost {
+                        debug!(
+                            address = %address,
+                            balance = %balance,
+                            "skipping drain, balance too low to cover gas"
+                        );
+                        return Ok::<_, BaselineError>(None);
+                    }
+
+                    let send_amount = balance.saturating_sub(drain_gas_cost);
+                    let wallet = EthereumWallet::from(signer);
+                    let provider = create_wallet_provider(primary_submission_rpc, wallet);
+                    let nonce = provider
+                        .get_transaction_count(address)
+                        .pending()
+                        .await
+                        .rpc("get pending transaction count")?;
+
+                    let tx = TransactionRequest::default()
+                        .with_to(funder_address)
+                        .with_value(send_amount)
+                        .with_nonce(nonce)
+                        .with_chain_id(chain_id)
+                        .with_gas_limit(drain_gas_limit as u64)
+                        .with_max_fee_per_gas(max_fee)
+                        .with_max_priority_fee_per_gas(max_priority_fee);
+
+                    match provider.send_transaction(tx).await {
+                        Ok(pending) => {
+                            let tx_hash = *pending.tx_hash();
+                            debug!(
+                                from = %address,
+                                amount = %send_amount,
+                                tx_hash = %tx_hash,
+                                "drain tx sent"
+                            );
+                            Ok(Some((address, send_amount)))
+                        }
+                        Err(e) => {
+                            warn!(from = %address, error = %e, "drain tx failed, skipping");
+                            Ok(None)
+                        }
+                    }
+                }
+            })
+            .collect();
+
+        let drain_results: Vec<_> = stream::iter(drain_futs)
+            .buffer_unordered(FUNDING_CONCURRENCY)
+            .inspect(|_| pb_drain.inc(1))
+            .collect()
+            .await;
+        pb_drain.finish_and_clear();
+
+        let mut pending_txs = Vec::new();
+        let mut total_drained = U256::ZERO;
+        for result in drain_results {
+            if let Some((address, amount)) = result? {
+                pending_txs.push(address);
+                total_drained = total_drained.saturating_add(amount);
+            }
+        }
+
+        if pending_txs.is_empty() {
+            info!("no accounts to drain");
+            return Ok(U256::ZERO);
+        }
+
+        let pb_confirm = self.progress_bar(pending_txs.len() as u64, "Waiting for drained funds");
+        info!(count = pending_txs.len(), total = %total_drained, "waiting for drained balances");
+
+        if let Err(e) =
+            Self::await_drained_balances(&client, &mut pending_txs, drain_gas_cost, &pb_confirm)
+                .await
+        {
+            warn!(error = %e, "some drain balances did not settle within timeout");
+        }
+        pb_confirm.finish_and_clear();
+
+        info!(total = %total_drained, "drain complete");
+        Ok(total_drained)
+    }
+
+    pub(super) fn progress_bar(&self, total: u64, prefix: &str) -> ProgressBar {
+        if self.snapshot_tx.is_some() {
+            return ProgressBar::hidden();
+        }
+        let pb = ProgressBar::new(total);
+        pb.set_style(
+            ProgressStyle::with_template("{prefix} [{bar:40.cyan/blue}] {pos}/{len} ({eta})")
+                .expect("valid template")
+                .progress_chars("█▓░"),
+        );
+        pb.set_prefix(prefix.to_string());
+        pb
+    }
+
+    /// Waits for account balances to reach a target after funding transfers.
+    async fn await_balances(
+        client: &QueryProvider,
+        pending_accounts: &mut Vec<Address>,
+        target_balance: U256,
+        pb: &ProgressBar,
+    ) -> Result<usize> {
+        let timeout = Duration::from_secs(60);
+        let poll_interval = Duration::from_millis(500);
+        let start = Instant::now();
+
+        let mut settled = 0usize;
+
+        while !pending_accounts.is_empty() && start.elapsed() < timeout {
+            tokio::time::sleep(poll_interval).await;
+
+            let mut still_pending = Vec::new();
+            for address in pending_accounts.drain(..) {
+                match client.get_balance(address).await.rpc("get balance") {
+                    Ok(balance) if balance >= target_balance => {
+                        debug!(address = %address, balance = %balance, "funding balance settled");
+                        settled += 1;
+                        pb.inc(1);
+                    }
+                    Ok(_) => {
+                        still_pending.push(address);
+                    }
+                    Err(e) => {
+                        warn!(address = %address, error = %e, "failed to check funding balance");
+                        still_pending.push(address);
+                    }
+                }
+            }
+            *pending_accounts = still_pending;
+        }
+
+        if !pending_accounts.is_empty() {
+            return Err(BaselineError::Transaction(format!(
+                "accounts did not reach funding target within timeout: {pending_accounts:?}"
+            )));
+        }
+
+        Ok(settled)
+    }
+
+
+    pub(super) async fn refresh_sender_state(&mut self) -> Result<()> {
+        let total_accounts = self.accounts.len();
+        let client = self.client.clone();
+        let pb_refresh = self.progress_bar(total_accounts as u64, "Refreshing account state");
+
+        let refresh_futs: Vec<_> = self
+            .accounts
+            .accounts()
+            .iter()
+            .map(|a| {
+                let client = client.clone();
+                let addr = a.address;
+                async move {
+                    let balance = client.get_balance(addr).await.rpc("get balance")?;
+                    let nonce =
+                        client.get_transaction_count(addr).await.rpc("get transaction count")?;
+                    Ok::<_, BaselineError>((addr, balance, nonce))
+                }
+            })
+            .collect();
+
+        let refresh_results: Vec<_> = stream::iter(refresh_futs)
+            .buffer_unordered(FUNDING_CONCURRENCY)
+            .inspect(|_| pb_refresh.inc(1))
+            .collect()
+            .await;
+        pb_refresh.finish_and_clear();
+
+        let addr_to_idx: HashMap<Address, usize> =
+            self.accounts.accounts().iter().enumerate().map(|(i, a)| (a.address, i)).collect();
+
+        let refresh_provider = RootProvider::<Ethereum>::new_http(self.config.query_rpc.clone());
+
+        for result in refresh_results {
+            let (addr, balance, account_nonce) = result?;
+            let idx = addr_to_idx[&addr];
+            let account = &mut self.accounts.accounts_mut()[idx];
+            account.balance = balance;
+            account.nonce = account_nonce;
+
+            let nonce_manager =
+                NonceManager::new(refresh_provider.clone(), addr, NONCE_RPC_TIMEOUT)
+                    .with_pending_tag();
+            Arc::make_mut(&mut self.nonce_managers).insert(addr, nonce_manager);
+
+            debug!(address = %addr, balance = %balance, nonce = account_nonce, "account state refreshed");
+        }
+
+        Ok(())
+    }
+
+    /// Waits for source account balances to drop to the post-drain dust threshold.
+    async fn await_drained_balances(
+        client: &QueryProvider,
+        pending_accounts: &mut Vec<Address>,
+        max_remaining: U256,
+        pb: &ProgressBar,
+    ) -> Result<usize> {
+        let timeout = Duration::from_secs(60);
+        let poll_interval = Duration::from_millis(500);
+        let start = Instant::now();
+        let mut settled = 0usize;
+
+        while !pending_accounts.is_empty() && start.elapsed() < timeout {
+            tokio::time::sleep(poll_interval).await;
+
+            let mut still_pending = Vec::new();
+            for address in pending_accounts.drain(..) {
+                match client.get_balance(address).await.rpc("get balance") {
+                    Ok(balance) if balance <= max_remaining => {
+                        debug!(address = %address, balance = %balance, "drain balance settled");
+                        settled += 1;
+                        pb.inc(1);
+                    }
+                    Ok(_) => {
+                        still_pending.push(address);
+                    }
+                    Err(e) => {
+                        warn!(address = %address, error = %e, "failed to check drain balance");
+                        still_pending.push(address);
+                    }
+                }
+            }
+            *pending_accounts = still_pending;
+        }
+
+        if !pending_accounts.is_empty() {
+            return Err(BaselineError::Transaction(format!(
+                "accounts did not drain within timeout: {pending_accounts:?}"
+            )));
+        }
+
+        Ok(settled)
+    }
+
+}
