+++ crates/infra/load-tests/src/workload/payloads/real_token_lifecycle.rs
@@ -1,3 +1,5 @@
+//! Real-token chain preparation for Uniswap/Aerodrome swap workloads.
+
 use std::{
     collections::BTreeSet,
     time::{Duration, Instant},
@@ -9,17 +11,35 @@ use alloy_provider::Provider;
 use alloy_rpc_types::{BlockNumberOrTag, TransactionRequest};
 use alloy_sol_types::{SolCall, sol};
 use futures::{StreamExt, stream};
-use indicatif::ProgressBar;
+use indicatif::{ProgressBar, ProgressStyle};
 use tracing::{debug, info, instrument, warn};
+use url::Url;
 
-use super::{
-    RealTokenAcquisition, RealTokenRecoverySummary, RealTokenSetup, SubmissionPipeline,
-    load_runner::{FUNDING_CONCURRENCY, LoadRunner},
-};
 use crate::{
     BaselineError, Result,
     rpc::{BaseFeeExt, QueryProvider, RpcResultExt, create_wallet_provider},
+    workload::{
+        AccountPool,
+        chain_prep::{
+            ChainPrepContext, RealTokenAcquisition, RealTokenRecoverySummary, RealTokenSetup,
+            await_token_balances, encode_erc20_balance_of, prep_submission_max_fee,
+        },
+    },
 };
+
+fn progress_bar(hide_progress: bool, total: u64, prefix: &str) -> ProgressBar {
+    if hide_progress {
+        return ProgressBar::hidden();
+    }
+    let pb = ProgressBar::new(total);
+    pb.set_style(
+        ProgressStyle::with_template("{prefix} [{bar:40.cyan/blue}] {pos}/{len} ({eta})")
+            .expect("valid template")
+            .progress_chars("█▓░"),
+    );
+    pb.set_prefix(prefix.to_string());
+    pb
+}
 
 const WETH_DEPOSIT_GAS_LIMIT: u64 = 100_000;
 const WETH_WITHDRAW_GAS_LIMIT: u64 = 45_000;
@@ -78,13 +98,12 @@ impl PairSwapDirection {
     }
 }
 
-impl LoadRunner {
-    fn collect_real_token_setup_approvals(
-        &self,
-        setup: &RealTokenSetup,
-    ) -> Vec<(Address, Address)> {
+fn collect_real_token_setup_approvals(
+    setup: &RealTokenSetup,
+    swap_routers: &[Address],
+) -> Vec<(Address, Address)> {
         let mut approvals = BTreeSet::new();
-        for router in self.collect_swap_routers() {
+        for &router in swap_routers {
             approvals.insert((setup.weth, router));
             approvals.insert((setup.pair_token.token, router));
         }
@@ -92,34 +111,37 @@ impl LoadRunner {
         approvals.into_iter().collect()
     }
 
-    /// Sets up real-token balances and approvals for bidirectional swap workloads.
-    ///
-    /// This preserves the existing random-direction swap payload semantics by
-    /// ensuring every sender has both WETH and the paired token before the
-    /// measured loop starts.
-    #[instrument(skip(self, setup), fields(accounts = self.accounts.len()))]
-    pub async fn setup_real_tokens(&mut self, setup: &RealTokenSetup) -> Result<()> {
-        let swap_routers = self.collect_swap_routers();
-        if swap_routers.is_empty() {
-            return Err(BaselineError::Config(
-                "real-token setup requires at least one swap router".into(),
-            ));
-        }
-        let approval_targets = self.collect_real_token_setup_approvals(setup);
+/// Sets up real-token balances and approvals for bidirectional swap workloads.
+///
+/// This preserves the existing random-direction swap payload semantics by
+/// ensuring every sender has both WETH and the paired token before the
+/// measured loop starts.
+#[instrument(skip(ctx, setup), fields(accounts = ctx.accounts.len()))]
+pub(super) async fn prepare(ctx: &mut ChainPrepContext<'_>, setup: &RealTokenSetup) -> Result<()> {
+    if ctx.outputs.real_tokens_prepared {
+        return Ok(());
+    }
+    let swap_routers = ctx.swap_routers.clone();
+    if swap_routers.is_empty() {
+        return Err(BaselineError::Config(
+            "real-token setup requires at least one swap router".into(),
+        ));
+    }
+    let approval_targets = collect_real_token_setup_approvals(setup, &swap_routers);
 
-        let base_fee = self.client.get_base_fee().await?;
+        let base_fee = ctx.client.get_base_fee().await?;
         let max_priority_fee = (base_fee / 10).max(1);
-        let max_fee = SubmissionPipeline::submission_max_fee(
+        let max_fee = prep_submission_max_fee(
             base_fee,
             max_priority_fee,
-            self.config.max_gas_price,
+            ctx.max_gas_price,
         );
 
         let account_data: Vec<_> =
-            self.accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
-        let client = self.client.clone();
-        let primary_submission_rpc = self.config.primary_submission_rpc().clone();
-        let chain_id = self.config.chain_id;
+            ctx.accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
+        let client = ctx.client.clone();
+        let primary_submission_rpc = ctx.primary_submission_rpc.clone();
+        let chain_id = ctx.chain_id;
         let setup = setup.clone();
         let total_accounts = account_data.len();
 
@@ -132,16 +154,16 @@ impl LoadRunner {
             "setting up real-token swap balances"
         );
 
-        let pb_setup = self.progress_bar(total_accounts as u64, "Setting up real tokens");
+        let pb_setup = ctx.progress_bar(total_accounts as u64, "Setting up real tokens");
         let setup_futs = account_data.into_iter().map(|(sender, signer)| {
             let client = client.clone();
             let primary_submission_rpc = primary_submission_rpc.clone();
             let setup = setup.clone();
             let approval_targets = approval_targets.clone();
             async move {
-                let weth_balance = Self::read_erc20_balance(&client, setup.weth, sender).await?;
+                let weth_balance = read_erc20_balance(&client, setup.weth, sender).await?;
                 let pair_balance =
-                    Self::read_erc20_balance(&client, setup.pair_token.token, sender).await?;
+                    read_erc20_balance(&client, setup.pair_token.token, sender).await?;
                 let needs_pair_token = pair_balance < setup.pair_token.amount_per_sender;
                 let acquisition_amount = if needs_pair_token {
                     setup.pair_token.acquisition.amount_in()
@@ -155,7 +177,7 @@ impl LoadRunner {
                 let mut approvals = Vec::new();
                 for (token, router) in &approval_targets {
                     let allowance =
-                        Self::read_erc20_allowance(&client, *token, sender, *router).await?;
+                        read_erc20_allowance(&client, *token, sender, *router).await?;
                     if allowance < setup.approval_amount {
                         approvals.push((*token, *router));
                     }
@@ -203,7 +225,7 @@ impl LoadRunner {
                     let tx = TransactionRequest::default()
                         .with_to(setup.weth)
                         .with_value(deposit_deficit)
-                        .with_input(Self::encode_weth_deposit())
+                        .with_input(encode_weth_deposit())
                         .with_nonce(nonce)
                         .with_chain_id(chain_id)
                         .with_gas_limit(WETH_DEPOSIT_GAS_LIMIT)
@@ -229,7 +251,7 @@ impl LoadRunner {
                 for (token, router) in approvals {
                     let tx = TransactionRequest::default()
                         .with_to(token)
-                        .with_input(Self::encode_erc20_approve(router, setup.approval_amount))
+                        .with_input(encode_erc20_approve(router, setup.approval_amount))
                         .with_nonce(nonce)
                         .with_chain_id(chain_id)
                         .with_gas_limit(ERC20_APPROVE_GAS_LIMIT)
@@ -254,7 +276,7 @@ impl LoadRunner {
                 }
 
                 if needs_pair_token {
-                    let (router, data) = Self::encode_pair_token_swap(
+                    let (router, data) = encode_pair_token_swap(
                         &setup,
                         sender,
                         PairSwapDirection::AcquirePairToken,
@@ -290,7 +312,7 @@ impl LoadRunner {
         });
 
         let setup_results: Vec<_> = stream::iter(setup_futs)
-            .buffer_unordered(FUNDING_CONCURRENCY)
+            .buffer_unordered(ctx.concurrency)
             .inspect(|_| pb_setup.inc(1))
             .collect()
             .await;
@@ -303,13 +325,13 @@ impl LoadRunner {
         }
 
         let sender_addresses: Vec<Address> =
-            self.accounts.accounts().iter().map(|a| a.address).collect();
+            ctx.accounts.accounts().iter().map(|a| a.address).collect();
 
         let mut weth_pending: Vec<_> =
             sender_addresses.iter().map(|sender| (setup.weth, *sender)).collect();
-        let pb_weth = self.progress_bar(weth_pending.len() as u64, "Waiting for WETH balances");
-        Self::await_token_balances(
-            &self.client,
+        let pb_weth = ctx.progress_bar(weth_pending.len() as u64, "Waiting for WETH balances");
+        await_token_balances(
+            ctx.client,
             &mut weth_pending,
             setup.weth_amount_per_sender,
             &pb_weth,
@@ -320,9 +342,9 @@ impl LoadRunner {
         let mut pair_pending: Vec<_> =
             sender_addresses.iter().map(|sender| (setup.pair_token.token, *sender)).collect();
         let pb_pair =
-            self.progress_bar(pair_pending.len() as u64, "Waiting for pair-token balances");
-        Self::await_token_balances(
-            &self.client,
+            ctx.progress_bar(pair_pending.len() as u64, "Waiting for pair-token balances");
+        await_token_balances(
+            ctx.client,
             &mut pair_pending,
             setup.pair_token.amount_per_sender,
             &pb_pair,
@@ -337,9 +359,9 @@ impl LoadRunner {
             }
         }
         let pb_allowance =
-            self.progress_bar(allowance_pending.len() as u64, "Waiting for router allowances");
-        Self::await_erc20_allowances(
-            &self.client,
+            ctx.progress_bar(allowance_pending.len() as u64, "Waiting for router allowances");
+        await_erc20_allowances(
+            ctx.client,
             &mut allowance_pending,
             setup.approval_amount,
             &pb_allowance,
@@ -347,7 +369,8 @@ impl LoadRunner {
         .await?;
         pb_allowance.finish_and_clear();
 
-        self.refresh_sender_state().await?;
+        ctx.outputs.needs_sender_refresh = true;
+        ctx.outputs.real_tokens_prepared = true;
 
         info!(
             accounts = total_accounts,
@@ -357,29 +380,31 @@ impl LoadRunner {
         Ok(())
     }
 
-    /// Recovers real-token balances by swapping the configured pair token back
-    /// into WETH, unwrapping WETH to native ETH, and leaving native drain to
-    /// [`Self::drain_accounts`].
-    pub async fn recover_real_tokens(
-        &self,
-        setup: &RealTokenSetup,
-    ) -> Result<RealTokenRecoverySummary> {
-        let client = self.client.clone();
-        let primary_submission_rpc = self.config.primary_submission_rpc().clone();
-        let chain_id = self.config.chain_id;
+/// Recovers real-token balances by swapping the configured pair token back
+/// into WETH, unwrapping WETH to native ETH, and leaving native drain to the runner.
+pub async fn recover_real_tokens(
+    client: &QueryProvider,
+    accounts: &AccountPool,
+    chain_id: u64,
+    max_gas_price: u128,
+    primary_submission_rpc: Url,
+    hide_progress: bool,
+    setup: &RealTokenSetup,
+) -> Result<RealTokenRecoverySummary> {
+        let client = client.clone();
 
         let base_fee = client.get_base_fee().await?;
         let max_priority_fee = (base_fee / 10).max(1);
-        let max_fee = SubmissionPipeline::submission_max_fee(
+        let max_fee = prep_submission_max_fee(
             base_fee,
             max_priority_fee,
-            self.config.max_gas_price,
+            max_gas_price,
         );
 
         let account_data: Vec<_> =
-            self.accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
+            accounts.accounts().iter().map(|a| (a.address, a.signer.clone())).collect();
         let total_accounts = account_data.len();
-        let pb_recover = self.progress_bar(total_accounts as u64, "Recovering real tokens");
+        let pb_recover = progress_bar(hide_progress, total_accounts as u64, "Recovering real tokens");
 
         let recover_futs: Vec<_> = account_data
             .into_iter()
@@ -398,10 +423,10 @@ impl LoadRunner {
                         .rpc("get pending transaction count")?;
 
                     let pair_balance =
-                        Self::read_erc20_balance(&client, setup.pair_token.token, sender).await?;
+                        read_erc20_balance(&client, setup.pair_token.token, sender).await?;
                     if pair_balance > U256::ZERO {
                         let recovery_router = setup.pair_token.acquisition.router();
-                        let allowance = Self::read_erc20_allowance(
+                        let allowance = read_erc20_allowance(
                             &client,
                             setup.pair_token.token,
                             sender,
@@ -413,7 +438,7 @@ impl LoadRunner {
                             let approval_amount = setup.approval_amount.max(pair_balance);
                             let tx = TransactionRequest::default()
                                 .with_to(setup.pair_token.token)
-                                .with_input(Self::encode_erc20_approve(
+                                .with_input(encode_erc20_approve(
                                     recovery_router,
                                     approval_amount,
                                 ))
@@ -441,7 +466,7 @@ impl LoadRunner {
                             nonce += 1;
                         }
 
-                        let (router, data) = Self::encode_pair_token_swap(
+                        let (router, data) = encode_pair_token_swap(
                             &setup,
                             sender,
                             PairSwapDirection::RecoverWeth,
@@ -480,11 +505,11 @@ impl LoadRunner {
                     }
 
                     let weth_balance =
-                        Self::read_erc20_balance(&client, setup.weth, sender).await?;
+                        read_erc20_balance(&client, setup.weth, sender).await?;
                     if weth_balance > U256::ZERO {
                         let tx = TransactionRequest::default()
                             .with_to(setup.weth)
-                            .with_input(Self::encode_weth_withdraw(weth_balance))
+                            .with_input(encode_weth_withdraw(weth_balance))
                             .with_nonce(nonce)
                             .with_chain_id(chain_id)
                             .with_gas_limit(WETH_WITHDRAW_GAS_LIMIT)
@@ -514,7 +539,7 @@ impl LoadRunner {
             .collect();
 
         let recover_results: Vec<_> = stream::iter(recover_futs)
-            .buffer_unordered(FUNDING_CONCURRENCY)
+            .buffer_unordered(crate::workload::chain_prep::PREP_CONCURRENCY)
             .inspect(|_| pb_recover.inc(1))
             .collect()
             .await;
@@ -537,7 +562,7 @@ impl LoadRunner {
         Ok(summary)
     }
 
-    async fn read_erc20_balance(
+async fn read_erc20_balance(
         client: &QueryProvider,
         token: Address,
         owner: Address,
@@ -546,7 +571,7 @@ impl LoadRunner {
             .call(
                 TransactionRequest::default()
                     .with_to(token)
-                    .with_input(Self::encode_erc20_balance_of(owner))
+                    .with_input(encode_erc20_balance_of(owner))
                     .into(),
             )
             .await
@@ -554,7 +579,7 @@ impl LoadRunner {
             .map(|bytes| U256::from_be_slice(bytes.as_ref()))
     }
 
-    async fn read_erc20_allowance(
+async fn read_erc20_allowance(
         client: &QueryProvider,
         token: Address,
         owner: Address,
@@ -564,7 +589,7 @@ impl LoadRunner {
             .call(
                 TransactionRequest::default()
                     .with_to(token)
-                    .with_input(Self::encode_erc20_allowance(owner, spender))
+                    .with_input(encode_erc20_allowance(owner, spender))
                     .into(),
             )
             .await
@@ -572,35 +597,35 @@ impl LoadRunner {
             .map(|bytes| U256::from_be_slice(bytes.as_ref()))
     }
 
-    fn encode_weth_deposit() -> Bytes {
+fn encode_weth_deposit() -> Bytes {
         sol! {
             function deposit() external payable;
         }
         Bytes::from(depositCall {}.abi_encode())
     }
 
-    fn encode_weth_withdraw(amount: U256) -> Bytes {
+fn encode_weth_withdraw(amount: U256) -> Bytes {
         sol! {
             function withdraw(uint256 wad) external;
         }
         Bytes::from(withdrawCall { wad: amount }.abi_encode())
     }
 
-    fn encode_erc20_approve(spender: Address, amount: U256) -> Bytes {
+fn encode_erc20_approve(spender: Address, amount: U256) -> Bytes {
         sol! {
             function approve(address spender, uint256 amount) external returns (bool);
         }
         Bytes::from(approveCall { spender, amount }.abi_encode())
     }
 
-    fn encode_erc20_allowance(owner: Address, spender: Address) -> Bytes {
+fn encode_erc20_allowance(owner: Address, spender: Address) -> Bytes {
         sol! {
             function allowance(address owner, address spender) external view returns (uint256);
         }
         Bytes::from(allowanceCall { owner, spender }.abi_encode())
     }
 
-    fn encode_pair_token_swap(
+fn encode_pair_token_swap(
         setup: &RealTokenSetup,
         recipient: Address,
         direction: PairSwapDirection,
@@ -646,8 +671,8 @@ impl LoadRunner {
         }
     }
 
-    /// Waits for ERC20 allowances to reach a target.
-    async fn await_erc20_allowances(
+/// Waits for ERC20 allowances to reach a target.
+async fn await_erc20_allowances(
         client: &QueryProvider,
         pending_allowances: &mut Vec<(Address, Address, Address)>,
         target_allowance: U256,
@@ -663,7 +688,7 @@ impl LoadRunner {
 
             let mut still_pending = Vec::new();
             for (token, owner, spender) in pending_allowances.drain(..) {
-                match Self::read_erc20_allowance(client, token, owner, spender).await {
+                match read_erc20_allowance(client, token, owner, spender).await {
                     Ok(allowance) if allowance >= target_allowance => {
                         debug!(token = %token, owner = %owner, spender = %spender, "allowance settled");
                         settled += 1;
@@ -695,4 +720,3 @@ impl LoadRunner {
 
         Ok(settled)
     }
-}
